requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.30
Alembic==1.13.1
Flask-Migrate==4.0.5
Marshmallow==3.21.2
Flask-Marshmallow==1.2.1
Flask-Login==0.6.3
python-dotenv==1.0.1
pytest==8.2.1
factory-boy==3.3.0
requests==2.32.3
google-generativeai==0.7.2

